services/chatbot/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ralph-stock-chatbot"
version = "0.1.0"
description = "Ralph Stock Chatbot - RAG-based Stock Analysis AI Chatbot"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Ralph Stock Team"}
]

dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "redis>=5.0.0",
    "google-generativeai>=0.3.0",
    # ralph-stock-lib는 Docker에서 로컬 경로로 설치됨
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "httpx>=0.24.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
markers = [
    "integration: marks tests as integration tests",
    "slow: marks tests as slow (docker build)",
]
